Add files via upload
Assignment 1: Taking input from user and printing output using only if statements.
take 3 user inputs, let's say a, b, & c and get their sum. Now, if a, b & c should >= 40, print as "Student is Passing" or print as "Student is Failing". 
and check if their is sum is >=125 and print as "Student is Passing" or print as "Student is Failing".
NOTE: Using if statements alone, no nested if or if else statements.